services/tph/requirements.txt
pytz
RPi.bme280
smbus2
pandas
redis
pytest
pytest-mock
